Fix/reported bugs

+++ models/item.py
@@ -11,7 +11,7 @@ class ItemModel(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(80), nullable=False)
     description = db.Column(db.String)
-    price = db.Column(db.Float(precision=2), nullable=False)
+    price = db.Column(db.Numeric(10, 2), nullable=False)
     image_url = db.Column(db.String(500), nullable=True)
     is_hidden = db.Column(db.Boolean, default=False, nullable=False)
